made bootstrap hero the default scaffolding
massively cleaned and reorganized the public folder. Also change quite
something in AppController and BaseController to get rid of the need to
have a related AppModel (which is not always needed) AND to be able to
hand over a specific Mako Template to render. Also fixed a bug in
TemplateLookup handling which is now working fine for /views/ and
/layouts/ directory.